fix(migrations): total_reviews mantém bigint em review_aggregates
CREATE OR REPLACE VIEW falha se mudar type de coluna. Em 0078 a coluna
total_reviews era count(*) cru = bigint. Eu tinha posto count(*)::int em
0084. Erro: 'cannot change data type of view column total_reviews from
bigint to integer'.

Removido o ::int — agora consistente com a assinatura original.
team_review_aggregates mantém ::int (em 0078 já era int).

+++ supabase/pending_migrations.sql
@@ -694,10 +694,12 @@ group by tr.reviewed_team_id;
 grant select on public.team_review_aggregates to authenticated, anon;
 
 -- Bónus: review_aggregates (singular, individual users) com mesmo filtro
+-- NOTA: total_reviews mantém-se bigint (sem ::int) para CREATE OR REPLACE
+-- não bater com o tipo declarado em 0078 (que era count(*) cru = bigint).
 create or replace view public.review_aggregates as
 select
   r.reviewed_id                                                                as user_id,
-  count(*)::int                                                                as total_reviews,
+  count(*)                                                                     as total_reviews,
   avg(coalesce(r.overall, (r.fair_play + r.punctuality + r.technical_level) / 3.0)) as avg_overall,
   avg(r.fair_play)                                                             as avg_fair_play,
   avg(r.punctuality)                                                           as avg_punctuality,
